Creates a "tas" namespace that can be used to call CMS2 methods via the extern'd CMS2 cms2 object without having to type cms2. everywhere.  All methods except for CMS2::Init and CMS2::GetEntry are included in the namespace.
Removed declaration of CMS2::ScanChain and removed the private sample histogram.
The macro no longer creates a file called CMS2.C, but rather a file called
ScanChain.C which defines only the int ScanChain method of the old CMS2.C.  This
is meant to serve as an example old looper upon which new old looper classes can
be built.  It uses the new tas namespace, but use of the namespace is not manda-
tory.

Updated several comments to reflect the current code structure, and provide up-to-
date instructions on how to compile and execute the code.