I have extended goodrun.cc so that it can directly read JSON files.  The semantics are as follows:
- If you want to read the good run list from a file called goodruns.json in your
  current working directory, simply say
          goodrun_json(run, lumi)
   analogously to how you've used goodrun(run, lumi) in the past.

- If your good-run list lives in a different json file, say
          set_goodrun_file_json("altgoodruns.json")
  once and thenceforth use goodrun_jsolumin, lumi)

Under the hood, the code calls Oli's conversion script.  If an error occurs
during the parsing of the converted list, the temp file that contains the
converted list will remain in your directory (same name as your json file, with
a .tmp appended) so you can examine what's wrong with it.

There are two known limitation with this code:

1.  The JSON parsing python library needs to be in your python path.
2.  You need write access to your current working directory so that the temp
    file with the plain-text run list can be created, and the temp file always
    has the same name.